3D Pong game w Joystick on seperate thread
This is the code for the completion of P. McWhorter's Course on Arduino with Python. The 'Final Exam' is to contribute something of your own to the code/tech that is shown in the course. i put the joystick control on it's own thread(cntrl_blocker function) to try to get the joystick to work better improving the interaction with Arduino and enhancing the playabiity of the game.